move the pandas ExtensionArray to its own module (issue #135)
