+++ mortie/__init__.py
@@ -2,6 +2,7 @@
 mortie: a library for generating morton indices
 """
 
+from importlib import import_module
 from importlib.metadata import PackageNotFoundError, version
 
 try:
@@ -123,9 +124,10 @@ __all__ = [
 # morton_index datatype (phase 5) + Arrow interop (phase 4) for issue #35. The
 # pandas ExtensionArray and the pyarrow ExtensionType are optional extras:
 # importing mortie must succeed with only numpy installed, so the names are
-# exposed lazily and built only when pandas / pyarrow are present (touching them
-# without the extra raises a clear ImportError). See mortie/morton_index.py and
-# mortie/arrow.py.
+# exposed lazily and resolved only when pandas / pyarrow are present (touching
+# them without the extra raises a clear ImportError). The ExtensionArray classes
+# themselves live in mortie/pandas.py, which mortie.morton_index imports on
+# demand (issue #135). See mortie/morton_index.py and mortie/arrow.py.
 from . import (
     arrow,  # noqa: F401
     morton_index,  # noqa: F401
@@ -150,6 +152,14 @@ _ARROW_NAMES = (
 
 
 def __getattr__(name):
+    if name == "pandas":
+        # mortie.morton_index imports this submodule eagerly when pandas is
+        # installed (to register the dtype string), which binds the attribute
+        # directly -- so this branch is reached only on a numpy-only install,
+        # where `mortie.pandas` must raise the curated ImportError rather than a
+        # bare AttributeError. `import_module`, not `from . import pandas`: the
+        # latter does a `hasattr` on this package first, re-entering __getattr__.
+        return import_module(f"{__name__}.pandas")
     if name in ("MortonIndexDtype", "MortonIndexArray"):
         return getattr(morton_index, name)
     if name in _ARROW_NAMES:
@@ -160,6 +170,10 @@ def __getattr__(name):
 __all__ += ['MortonIndexDtype', 'MortonIndexArray', 'morton_index']
 __all__ += ['decimal_to_word', 'decimals_to_words']
 __all__ += list(_ARROW_NAMES) + ['arrow']
+# 'pandas' is deliberately NOT in __all__, unlike the 'morton_index' / 'arrow'
+# submodules: `from mortie import *` would then bind the name `pandas` to
+# mortie's submodule in the caller's namespace, shadowing the real pandas there.
+# `import mortie.pandas` / `mortie.pandas` reach it explicitly (issue #135).
 
 # The Rust extension is imported and used internally by the tools.py encoders
 # No need to do anything here - tools.py handles the Rust integration
